requirements.txt
fastapi
uvicorn[standard]
feedparser
anthropic
python-dotenv
